+++ test/unit/auth/credential.spec.ts
@@ -31,8 +31,8 @@ import * as utils from '../utils';
 import * as mocks from '../../resources/mocks';
 
 import {
-  ApplicationDefaultCredential, CertCredential, Certificate, GoogleOAuthAccessToken,
-  MetadataServiceCredential, RefreshTokenCredential,
+  ApplicationDefaultCredential, CertCredential, Certificate, Credential, GoogleOAuthAccessToken,
+  MetadataServiceCredential, RefreshToken, RefreshTokenCredential,
 } from '../../../src/auth/credential';
 import { HttpClient } from '../../../src/utils/api-request';
 import {Agent} from 'https';
@@ -337,7 +337,7 @@ describe('Credential', () => {
       if (fsStub) {
         fsStub.restore();
       }
-      process.env.GOOGLE_APPLICATION_CREDENTIALS = this.credPath;
+      process.env.GOOGLE_APPLICATION_CREDENTIALS = credPath;
     });
 
     it('should return a CertCredential with GOOGLE_APPLICATION_CREDENTIALS set', () => {
@@ -354,6 +354,19 @@ describe('Credential', () => {
     it('should throw if explicitly pointing to an invalid cert file', () => {
       fsStub = sinon.stub(fs, 'readFileSync').returns('invalidjson');
       expect(() => new ApplicationDefaultCredential()).to.throw(Error);
+    });
+
+    it('should throw error if type not specified on cert file', () => {
+      fsStub = sinon.stub(fs, 'readFileSync').returns(JSON.stringify({}));
+      expect(() => new ApplicationDefaultCredential())
+          .to.throw(Error, 'Invalid contents in the credentials file');
+    });
+
+    it('should throw error if type is unknown on cert file', () => {
+      fsStub = sinon.stub(fs, 'readFileSync').returns(JSON.stringify({
+        type: 'foo',
+      }));
+      expect(() => new ApplicationDefaultCredential()).to.throw(Error, 'Invalid contents in the credentials file');
     });
 
     it('should return a RefreshTokenCredential with gcloud login', () => {
@@ -394,6 +407,24 @@ describe('Credential', () => {
         clientEmail: mockCertificateObject.client_email,
         privateKey: mockCertificateObject.private_key,
       });
+    });
+
+    it('should parse valid RefreshTokenCredential if GOOGLE_APPLICATION_CREDENTIALS environment variable ' +
+        'points to default refresh token location', () => {
+      process.env.GOOGLE_APPLICATION_CREDENTIALS = GCLOUD_CREDENTIAL_PATH;
+
+      fsStub = sinon.stub(fs, 'readFileSync').returns(JSON.stringify(MOCK_REFRESH_TOKEN_CONFIG));
+
+      const adc = new ApplicationDefaultCredential();
+      const c = adc.getCredential();
+      expect(c).is.instanceOf(RefreshTokenCredential);
+      expect(c).to.have.property('refreshToken').that.includes({
+        clientId: MOCK_REFRESH_TOKEN_CONFIG.client_id,
+        clientSecret: MOCK_REFRESH_TOKEN_CONFIG.client_secret,
+        refreshToken: MOCK_REFRESH_TOKEN_CONFIG.refresh_token,
+        type: MOCK_REFRESH_TOKEN_CONFIG.type,
+      });
+      expect(fsStub.alwaysCalledWith(GCLOUD_CREDENTIAL_PATH, 'utf8')).to.be.true;
     });
   });