+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - [fixed] Implemented a Node.js environment check that will be executed at
   package import time.
+- [fixed] Setting GOOGLE_APPLICATION_CREDENTIALS environment variable  
+  to a refresh token instead of a certificate token now supported
 
 # v6.5.0